src/functions/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.4.6
    # via aiohttp
aiohttp==3.11.12
    # via azure-functions-durable
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via httpx
assemblyai==0.37.0
    # via functions (pyproject.toml)
attrs==25.1.0
    # via aiohttp
azure-core==1.32.0
    # via
    #   azure-identity
    #   azure-storage-blob
azure-functions==1.21.3
    # via
    #   functions (pyproject.toml)
    #   azure-functions-durable
azure-functions-durable==1.2.10
    # via functions (pyproject.toml)
azure-identity==1.19.0
    # via functions (pyproject.toml)
azure-storage-blob==12.24.1
    # via functions (pyproject.toml)
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
cryptography==44.0.0
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
furl==2.1.3
    # via azure-functions-durable
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via assemblyai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
isodate==0.7.2
    # via azure-storage-blob
msal==1.31.1
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.2.0
    # via azure-identity
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
orderedmultidict==1.0.1
    # via furl
portalocker==2.10.1
    # via msal-extensions
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
pycparser==2.22
    # via cffi
pydantic==2.10.6
    # via assemblyai
pydantic-core==2.27.2
    # via pydantic
pyjwt==2.10.1
    # via msal
python-dateutil==2.9.0.post0
    # via azure-functions-durable
requests==2.32.3
    # via
    #   azure-core
    #   azure-functions-durable
    #   msal
six==1.17.0
    # via
    #   azure-core
    #   furl
    #   orderedmultidict
    #   python-dateutil
sniffio==1.3.1
    # via anyio
typing-extensions==4.12.2
    # via
    #   anyio
    #   assemblyai
    #   azure-core
    #   azure-identity
    #   azure-storage-blob
    #   pydantic
    #   pydantic-core
urllib3==2.3.0
    # via requests
websockets==14.2
    # via assemblyai
yarl==1.18.3
    # via aiohttp
